Move trace printing to getDependentStoryFiles
